requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai==1.88.0
gunicorn==21.2.0
flask-limiter==3.5.0
flask-cors==4.0.0 